requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.11
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib-metadata==7.0.2
jiter==0.9.0
lxml==5.2.2
multidict==6.1.0
numpy>=1.26.0,<2.0
openai==1.66.3
openpyxl==3.1.2
pandas==2.1.4
pdfplumber==0.10.3
Pillow==10.3.0   # Added for PIL
geopy==2.4.1     # Added for geolocation
matplotlib==3.8.2  # Added for visualization
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
python-multipart==0.0.20
requests==2.32.3
sniffio==1.3.1
starlette==0.46.1
tiktoken==0.5.2  # Added for OpenAI token counting
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
yarl==1.18.3
yt_dlp==2024.03.10
zipfile36==0.1.3
